rules pages separated into its own component
